+++ src/ch/alexi/fractgen/models/FractCalcerResultData.java
@@ -12,7 +12,7 @@ import java.awt.image.BufferedImage;
 public class FractCalcerResultData {
 	public BufferedImage fractImage;
 	public FractParam fractParam;
-	public int[][] iterValues;
+	public double[][] iterValues;
 	public RGB[] colorPalette;
 	
 	public FractCalcerResultData() {
@@ -22,7 +22,7 @@ public class FractCalcerResultData {
 	public FractCalcerResultData(FractParam fractParam, BufferedImage img,RGB[] colorPalette) {
 		this.fractParam = fractParam;
 		this.fractImage = img;
-		this.iterValues = new int[fractParam.picWidth][fractParam.picHeight];
+		this.iterValues = new double[fractParam.picWidth][fractParam.picHeight];
 		this.colorPalette = colorPalette;
 	}
 }
